refactor: Generate code to register host functions in wasm vm (#8176)

+++ src/tests/libxrpl/tx/wasm/host_context/OracleKeylet.cpp
@@ -17,14 +17,13 @@ struct OracleKeyletCall : HostContextTest
     Bytes const accountBytes{0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0a,
                              0x0b, 0x0c, 0x0d, 0x0e, 0x0f, 0x10, 0x11, 0x12, 0x13, 0x14};
     AccountID const account = AccountID::fromVoid(accountBytes.data());
-    std::int32_t const docId = 12345;
+    std::uint32_t const docId = 12345;
 };
 
 TEST_F(OracleKeyletCall, AccountAndDocIdAreForwardedKeyletIsWritten)
 {
     Bytes const keylet(32, 0xab);
-    EXPECT_CALL(host, oracleKeylet(account, static_cast<std::uint32_t>(docId)))
-        .WillOnce(testing::Return(keylet));
+    EXPECT_CALL(host, oracleKeylet(account, docId)).WillOnce(testing::Return(keylet));
 
     OutRegion out{32};
     EXPECT_EQ(
@@ -35,7 +34,7 @@ TEST_F(OracleKeyletCall, AccountAndDocIdAreForwardedKeyletIsWritten)
 
 TEST_F(OracleKeyletCall, HostErrorBecomesContractReturnValue)
 {
-    EXPECT_CALL(host, oracleKeylet(account, static_cast<std::uint32_t>(docId)))
+    EXPECT_CALL(host, oracleKeylet(account, docId))
         .WillOnce(testing::Return(std::unexpected(HostFunctionError::LedgerObjNotFound)));
 
     OutRegion out{32};
@@ -79,7 +78,7 @@ TEST_F(OracleKeyletCall, EmptyAccountIsRefusedWithoutAskingHost)
 
 TEST_F(OracleKeyletCall, HostExceptionBecomesInternalFatalAndIsLogged)
 {
-    EXPECT_CALL(host, oracleKeylet(account, static_cast<std::uint32_t>(docId)))
+    EXPECT_CALL(host, oracleKeylet(account, docId))
         .WillOnce(testing::Throw(std::runtime_error{"oracle keylet came apart"}));
 
     OutRegion out{32};
@@ -95,8 +94,7 @@ TEST_F(OracleKeyletCall, HostExceptionBecomesInternalFatalAndIsLogged)
 TEST_F(OracleKeyletCall, ShortOutRegionWritesNothingAndReturnsTrueLength)
 {
     Bytes const keylet(32, 0xab);
-    EXPECT_CALL(host, oracleKeylet(account, static_cast<std::uint32_t>(docId)))
-        .WillOnce(testing::Return(keylet));
+    EXPECT_CALL(host, oracleKeylet(account, docId)).WillOnce(testing::Return(keylet));
 
     OutRegion out{keylet.size() - 1};
     EXPECT_EQ(
@@ -108,8 +106,7 @@ TEST_F(OracleKeyletCall, ShortOutRegionWritesNothingAndReturnsTrueLength)
 TEST_F(OracleKeyletCall, OutRegionOfExactSizeIsWritten)
 {
     Bytes const keylet(32, 0xab);
-    EXPECT_CALL(host, oracleKeylet(account, static_cast<std::uint32_t>(docId)))
-        .WillOnce(testing::Return(keylet));
+    EXPECT_CALL(host, oracleKeylet(account, docId)).WillOnce(testing::Return(keylet));
 
     OutRegion out{keylet.size()};
     EXPECT_EQ(
@@ -120,8 +117,7 @@ TEST_F(OracleKeyletCall, OutRegionOfExactSizeIsWritten)
 
 TEST_F(OracleKeyletCall, EmptyResultAnswersZeroAndWritesNothing)
 {
-    EXPECT_CALL(host, oracleKeylet(account, static_cast<std::uint32_t>(docId)))
-        .WillOnce(testing::Return(Bytes{}));
+    EXPECT_CALL(host, oracleKeylet(account, docId)).WillOnce(testing::Return(Bytes{}));
 
     OutRegion out{32};
     EXPECT_EQ(hostContext.oracleKeylet(bytesOf(accountBytes), docId, out.slice()), 0);
